requirements.txt
# To install these packages, run: pip install -r requirements.txt

Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-SocketIO==5.3.3
Flask-Cors==3.0.10
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
python-dotenv==1.0.0
SQLAlchemy==1.4.46
